pyproject.toml
[tool.poetry]
name = "test-code"
version = "0.1.0"
description = ""
authors = ["kokorii <kokori@kakao.com>"]
readme = "README.md"
packages = [{include = "test_code"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest-playwright = "^0.3.3"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
